+++ test/_todo.js
@@ -1,7 +1,9 @@
 describe('TODO', function () {
 
-  it('has opts to stop');
-
-  it('uses emit instead of send');
+  it('remove scriptPath from .stop and .kill');
+  it('added opts to stop and kill');
+  it('added opts.timeout');
+  it('send to child with multiple event args');
+  it('emit instead of send for ipc');
 
 });