+++ test/with-child.js
@@ -6,15 +6,15 @@ describe('with child', function () {
 
   var scriptFile = path.resolve(__dirname, 'procs', 'server-with-non-errors');
 
-  var hookRef = MochaFork.before.start(scriptFile);
+  var childRef = MochaFork.before.start(scriptFile);
 
-  hookRef.after.stop();
+  childRef.after.stop();
 
   it('has access to child', function () {
 
-    expect(typeof hookRef._child).to.be('object');
+    expect(typeof childRef._child).to.be('object');
 
-    expect(typeof hookRef._child.pid).to.be('number');
+    expect(typeof childRef._child.pid).to.be('number');
 
   });