+++ test/with-message-to-child.js
@@ -6,13 +6,13 @@ describe('with message to child', function () {
 
   var scriptFile = path.resolve(__dirname, 'procs', 'server-receive-message-from-parent');
 
-  var hookRef = MochaFork.before.start(scriptFile);
+  var childRef = MochaFork.before.start(scriptFile);
 
   it ('can send message to child', function (done) {
 
-    hookRef.send('event-name', {kill: true});
+    childRef.send('event-name', {kill: true});
 
-    hookRef.on('exit', function () {
+    childRef.on('exit', function () {
 
       done();