+++ test/with-not-error.js
@@ -6,9 +6,9 @@ describe('with not error', function () {
 
   var scriptFile = path.resolve(__dirname, 'procs', 'server-with-non-errors');
 
-  var hookRef = MochaFork.before.start(scriptFile);
+  var childRef = MochaFork.before.start(scriptFile);
 
-  hookRef.after.stop();
+  childRef.after.stop();
 
   it('started and stopped', function () {});