+++ test/with-stop.js
@@ -11,9 +11,9 @@ describe('with stop', function () {
     host: 'localhost'
   };
 
-  var hookRef = MochaFork.before.start(scriptFile, scriptOpts);
+  var childRef = MochaFork.before.start(scriptFile, scriptOpts);
 
-  hookRef.after.stop();
+  childRef.after.stop();
 
   it('started server', function (done) {